+++ src/main/java/power/keepeersofthestones/procedures/ClusterSharpIsGoneProcedure.java
@@ -31,7 +31,7 @@ public class ClusterSharpIsGoneProcedure {
 			}
 
 			private void run() {
-				world.setBlock(new BlockPos((int) x, (int) y, (int) z), Blocks.AIR.defaultBlockState(), 3);
+				world.setBlock(new BlockPos(x, y, z), Blocks.AIR.defaultBlockState(), 3);
 				MinecraftForge.EVENT_BUS.unregister(this);
 			}
 		}.start(world, 400);
