+++ src/main/java/power/keepeersofthestones/procedures/EmitateSkeletonProcedure.java
@@ -16,8 +16,8 @@ public class EmitateSkeletonProcedure {
 			return;
 		if (world instanceof Level _level) {
 			if (!_level.isClientSide()) {
-				_level.playSound(null, new BlockPos((int) x, (int) y, (int) z),
-						ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("entity.skeleton.ambient")), SoundSource.HOSTILE, 1, 1);
+				_level.playSound(null, new BlockPos(x, y, z), ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("entity.skeleton.ambient")),
+						SoundSource.HOSTILE, 1, 1);
 			} else {
 				_level.playLocalSound(x, y, z, ForgeRegistries.SOUND_EVENTS.getValue(new ResourceLocation("entity.skeleton.ambient")),
 						SoundSource.HOSTILE, 1, 1, false);
