+++ src/main/java/power/keepeersofthestones/procedures/JoinPlayerInTheWorldProcedure.java
@@ -45,7 +45,7 @@ public class JoinPlayerInTheWorldProcedure {
 					.orElse(new PowerModVariables.PlayerVariables())).selected) {
 				{
 					if (entity instanceof ServerPlayer _ent) {
-						BlockPos _bpos = new BlockPos((int) x, (int) y, (int) z);
+						BlockPos _bpos = new BlockPos(x, y, z);
 						NetworkHooks.openGui((ServerPlayer) _ent, new MenuProvider() {
 							@Override
 							public Component getDisplayName() {
